Add Player and Food classes
Created scripts for the Player and Food Blob classes. Imported
BayatGames's free asset package in order to test the autogeneration of
Blob Sprite icons upon loading a Blob Food or Player object. Finally,
create Prefabs for the Food and Player objects.
Note that the Prefabs as well as the Blob classes are unfinished.