Make value depend on action
Modified the classes to that the Consumable's FoodValue instance
variable depends on its ConsumableAction's FoodValue, instead of being a
separate value from its action. This way, if the action's FoodValue
changes, then we no longer have to update the FoodValue in the
Consumable object as well.

Because of this update, the FoodValue of a Blob had to be made an
abstract auto-property, which is overridden in the Consumable and Player
sub-classes.