Calculate velocity changes upon collision.
Calculated the changes in velocity between two colliding objects using
f = m * a, where the x-coordinate of the velocity changes by f *
cos(theta) and the y-coordinate of the velocity changes by f *
sin(theta). The angle theta is calculated as the angle between the
velocity of the moving object and the positive x-axis.