Fix collision acceleration calculation
Changed the collision function when detecting collision between two Blob
objects. Previously, in the function f = m * a, we used the mass of the
collided Blob with the acceleration of the current Blob that was
detecting the collision. Now, we changed it to be the mass and
acceleration of the collided Blob, as it should be.
Also added conditional statement to check if the collided object is a
player or a food object.