Create auto-instantiator for food
Created a prefab and class that will auto instantiate food objects for
the given level.
Modified the Consumable objects so that they start larger and grow at a
smaller rate for objects with a higher FoodValue. This way, we don't get
very small or very large food objects, but instead we get a more
consistent size for them.