Modify collision event
Modified the collision events. Moved some functionality to Consumable in
order to begin work on using Strategy for player's actions to change
based on powerups.

The player will have an Action object, which will contain the
ConsumeFood(), OnConsumeFoodEvent(), ConsumePowerUp(), and
OnConsumePowerUpEvent() methods. When the Player has a collision event
with a Food or PowerUp, we will call player.action.ConsumeFood(food) or
player.action.ConsumePowerUp(powerup) for food and powerups
respectively. The Shrink() Coroutine found in the Consumable will stay,
and if a Player consumes a powerup, we will not stop the Shrinking. This
will be done after the canvas is made.