Create action for players
Finished creating the PlayerAction and DefaultPlayerAction classes so
that a Player can have dynamic functionality (using strategy) based on
which PlayerAction is currently attached to the Player object. This way,
we can easily implement powerups, and just create a new action to attach
to the player for when it injests a particular power up.