Add knight theme and create global theme class
Created a static class which controls which theme is currently enabled.
The game driver consults this class when theming the game, and the menu
should also consult this class when deciding what the main menu should
look like.

Added a new Knight theme with baddies as negative consumables and
potions as positive consumables.