Fix audio bugs
Fixed a bug where the sound toggle would not work when switching from
one scene then back to the main menu.
Fixed bug where sound icon would display "sound" when switching between
scenes even when the audio was muted. When muting the sound then
switching to a new scene and then back to the main menu, the sound icon
would infer that sound was playing when the icon should really have been
the mute icon. This bug has been fixed.